Add remark command skeletal implementation
The current AddressBook does not support adding additional
information about a person.

Allowing the user to add remarks about a person in the
AddressBook can let them record down important
additional information about someone so that they
can better organize and keep track of people of
interest.

Let's add a remark command to the AddressBook.

The current implementation is not yet complete and only
a partial one. The complete one will be done very soon.