Add Age and Gender Attributes
Add Age and Gender Attributes with Testing to Patient

There is important information, such as the age and gender, of patients
that should be known to users. Such information is missing in the code.

Adding the age and gender attributes ensures that we have important
details of patients.

Let’s create the classes: Age, Gender, AgeTest and GenderTest. Doing so
will ensure that the user can input important information about patients
into the application and ease subsequent testing of the information.