Add AssignedDepartment attribute to Patient
The Patient class does not currently store information
regarding the department the patient is assigned to.

Adding an attribute that stores that information can
help us create the starting ground to implement the
functionality of assigning patients to departments.

Let's,
* Create the AssignedDepartment class, which is a
  Patient attribute
* Create the Department Enumeration, which represents
  the departments in the hospital

Using an enumeration to represent the departments will
allow us to constrain the department information to a
limited predefined set of values. This implementation
might however be changed in the future should the need
arises (i.e. if we want the user to be able to customize
the available departments, or add more functionality to
the department type).