Fix inconsistencies in EditCommand and EditCommandParser
Currently, the EditCommand and EditCommandParser classes are
inconsistent in that they contain methods that process
patient attributes that are not parameters into the Edit
command, or certain parts lack the processing of important
attributes.

Fixing this issue will prevent further bugs and lead to
neater code.

Let's,
* Add and remove attributes in the two classes for
consistency
* Make sure that important attributes are mentioned in
parts that were previously lacking them