Fix StackOverflow bug involving PatientBuilder and RecordBuilder
There is an issue where after PatientBuilder() is called,
it will call RecordBuilder(), which in turn calls
PatientBuilder() again, causing a StackOverflow error.

We need to fix this to get our test cases to pass.

Let's
* Add a new constructor to RecordBuilder that takes in
a PatientBuilder to initialize its patient, instead
of calling PatientBuilder again
* Add new attributes to TypicalPatients

Note that not all of the fields in Typical Patients are
updated! AMY and BOB still need to be updated, which
involves adding new attributes into CommandTestUtilgit add --all