Tweak layout of GUI to resemble our ui design
The GUI currently only displays the patient list in
the entire window.

We need to update the layout of the GUI so that it
more closely resembles the final product.

Let's change the layout of the main window to be more
similar to our proposed UI design. Let's also add
the new attributes to each patient so that the user
can refer to them when they scan the list.

Note that this layout is still tentative and subject
to change.