Modify Delete Command to use IC Number
The existing delete command uses the index to delete a patient from the list. This can be very cumbersome especially in very large datasets of patients.

Hence, use IC Number of the patient instead to delete a patient directly from the list like "delete T1234567B"

Let's update the Delete Command and fix tests for the DeleteCommand and Parser tests.