Modify Edit Command to use IC Number
The existing edit command uses the index to edit a patient's description in the list. This can be very cumbersome especially in very large datasets of patients.

Hence, use IC Number of the patient instead to edit a patient's description directly from the list like "edit T1234567B n/NEW NAME"

Let's update the EDIT Command and fix tests for the EditCommand and Parser tests.