Update UI Component section of developer guide
Currently the UI Component section of the developer guide
is outdated and missing information related to the
newly added RecordCard and RecordPanel classes.

Updating the guide can help us keep better documentation
of our project.

Let's add the newly added RecordPanel and RecordCard
classes to the UML diagram and alter the explanation
to include them as well.