Update DG with Implementation of Record Feature
Update DG with Implementation of Record Command

A New Record Command has been created and implemented. DG does
not contain information about the Record Command.

Updating the DG will ensure that DG is kept up to date and fellow
developers understand the implementation of the newly added Record command.

Let’s keep the DG up to date with our enhancements. Doing so allows
everyone to be aware of the changes made and how they were made.