Create Method to Copy Record Information
Create Method to Copy Record Information

The Patient class has a constructor that takes in Record as a
parameter and sets its record field to the Record given in the
parameter. However, this would make a patient have a Record
that could be tied to another Patient.

Creating a method to copy the Record information will ensure that
the Patient field of the Record matches.

Let’s be careful in our code. Doing so will allow our product to be
more future-proof and error-free.