Fix bug where patient details do not update
Currently, when you edit a patient's details, the GUI
does not update with the new details unless you reselect
the patient in the list.

We must fix this bug so that the user can view the new
details easily.

Let's edit the RecordPanel class to handle the event
when the patient's details changes properly.