Restyle Record fields and fix non-update bug
The record fields in the GUI currently do not update
when the user edits the record using the record
command. There is also the remark field in the GUI
even though we do not intend for the user to edit it.

We should fix the bug and remove the remark so that the
program works like intended.

Let's
* Wrap the Record fields in SimpleStringProperty so that
changes to them can be observed by listeners in the GUI.
* Change the container type of record fields in the GUI
from TextFlow to TextArea.
* Add the GUI components as listeners to the events of the
Record Card fields.
* Remove the remark box from the GUI."