Simplify the previous changes
The changes needed for the GUI to update when record
is edited previously involved the refactoring of
String attributes in Record into SimpleStringProperty.

However, such changes in Record were not required for
the bug fix we need for the GUI.

Let's simplify the changes so that we can avoid
modifying what we have previously done as much
as possible.