Make final changes and fix warnings
When running the application, there are many warnings about
the wrong JavaFX version that show up in the console.

We should remove these warnings so that the application can
run as smoothly as possible.

Let's fix these warnings and also make some final changes
to the GUI of the application.