Add Priority Attribute
Add Priority Attribute

Having priority as an attribute of patients would be helpful for ED
doctors. Our system does not have a Priority attribute.

Creating a class for Priority will ensure that this can be implemented
as part of a Patient.

Let’s create the classes: Priority and PriorityTest. Doing so
will ensure that the user can input important information about patients
into the application and ease subsequent testing of the information.