Implement Priority Attribute
Implement Priority Attribute

Priority attribute has been created for our Patient class. However, it
is not being implemented in our system.

Editing the relevant classes will ensure that the Priority attribute is
implemented and functional.

Let’s edit the Patient, Command and etc. classes to implement
Priority. Doing so will ensure that the Priority attribute can be used
and provide benefits to users.