Fix Patient test case and relax equals for Patient
Currently, there is an issue with the test case for equals()
in Patient. This test case is believed to arise from the
addition of the age.equals into the equals method.

Perhaps removing it will fix the test case failure.

Let's remove the equality checking for age in Patient and
add age into the toString test case expected output."