Update test cases for Age and Birthday
Currently, the AgeTest tests and BirthdayTests are
outdated and do not take into account the new changes.

We should update the test cases to ensure that our
tests are testing the updated methods properly.

Let's add new test cases for the constructor and toString
methods in Age, and for the isValidBirthdate method in
BirthdayTest.