Add sort command to app
Currently, there lacks a command for users to sort
the patient list by a given property.

We should add it to allow patients to better
organize their data.

Let's,
* Add the SortCommand and SortCommandParser
classes
* Modify Priority class to store the enum type
as its value instead of a string

Note that the function is currently implemented
such that you can sort the displayed filtered list,
but doing so will also sort the underlying full list.
It is this way as the filteredpatientlist cannot be
modified, so sorting the underlying addressbook is
the fastest and easiest way to add this functionality
without modifying the structure of the model and ui
too much."