Make final UI changes
Currently, the styling of the scrollbar looks out
of place with the rest of the UI. The help window's
font is different as well. Warnings are also thrown
when FXML files are loaded.

We should update the UI and remove these warnings
so that our GUI looks more cohesive and runs more
smoothly.

Let's,
* Style the scrollbar
* Style the help window
* Remove fxml warnings