Update images in user guide
Right now, the image for checking Java version is
missing. The image for the jar file is blurry as well.

We should replace these images so that users will find
the images helpful.

Let's add the java version image and replace the jar file
image with a high quality one.