Fix name sorting by ASCII value
Currently, the sorting of names follow the ASCII
order, whereby all names starting with a capital
letter will be ordered in front of names starting
with a lower case letter.

However, this does not make much sense from the
user's standpoint, as we would expect sorting by
lexicographical order to be based purely on the
alphabet and not the case.

Let's change the compareTo method in name to
disregard the case of letters when comparing.