Fix error message bugs
Currently, if you use the assign command without department
prefix, the program will not display an error message but an
exception is thrown behind the scenes.

This is undesirable as the user does not get any feedback
as to why the command is not working, and there will
be an unhandled exception which might cause other issues
in the program.

Let's add a check for department prefix in AssignCommandParser,
and also add a newline to the assign command message.