Update DG with Appendix Instructions for Manual Testing
Update DG with Appendix Instructions for Manual Testing

An important appendix for manual testing is missing in the DG.

Editing the DG ensures that the DG is more complete and relevant.

Let’s keep the DG updated. Doing so ensures future and fellow
developers have a clearer idea of our project.