Update DG
Update DG

Implementation of Record Command and Appendix: Instructions for Manual
Testing in DG are not updated and polished.

Editing the DG ensures that our DG is more complete for submission.

Let’s update our parts in the DG. Doing so will allow us to get better marks
for DG.